Refactor torch example .py to examples folder in beam
