+++ sdks/python/apache_beam/examples/inference/pytorch_image_classification.py
